refactor: move onboarding helpers from types.ts to config.ts
Break circular import cycle between types.ts and config.ts (#519).
resolveDingTalkAccount and listDingTalkAccountIds are config resolution
functions that belong in config.ts alongside getConfig, mergeAccountWithDefaults,
and isConfigured.
