+++ src/types.ts
@@ -19,7 +19,6 @@ import type {
   ChannelLogSink as SDKChannelLogSink,
 } from "openclaw/plugin-sdk/channel-runtime";
 import type { ChannelSetupWizard } from "openclaw/plugin-sdk/setup";
-import { mergeAccountWithDefaults } from "./config";
 
 export type AckReactionMode = "off" | "emoji" | "kaomoji";
 // Accept arbitrary strings for backward compatibility; the recommended
@@ -732,149 +731,3 @@ export interface ConnectionAttemptResult {
   nextDelay?: number;
 }
 
-// ============ Onboarding Helper Functions ============
-
-const DEFAULT_ACCOUNT_ID = "default";
-
-function stripRemovedLegacyFieldsFromPublicAccount(
-  config: DingTalkConfig,
-): DingTalkConfig {
-  const {
-    cardStreamReasoning: _cardStreamReasoning,
-    verboseRealtimeStream: _verboseRealtimeStream,
-    accounts,
-    ...rest
-  } = config as DingTalkConfig & {
-    cardStreamReasoning?: unknown;
-    verboseRealtimeStream?: unknown;
-    accounts?: Record<string, DingTalkConfig | undefined>;
-  };
-  const sanitizedAccounts = accounts
-    ? Object.fromEntries(
-        Object.entries(accounts).map(([accountId, accountConfig]) => [
-          accountId,
-          accountConfig ? stripRemovedLegacyFieldsFromPublicAccount(accountConfig) : accountConfig,
-        ]),
-      )
-    : undefined;
-  if (sanitizedAccounts) {
-    return { ...rest, accounts: sanitizedAccounts } as DingTalkConfig;
-  }
-  return rest as DingTalkConfig;
-}
-
-/**
- * List all DingTalk account IDs from config
- */
-export function listDingTalkAccountIds(cfg: OpenClawConfig): string[] {
-  const dingtalk = cfg.channels?.dingtalk as DingTalkChannelConfig | undefined;
-  if (!dingtalk) {
-    return [];
-  }
-
-  const accountIds: string[] = [];
-
-  // Check for direct configuration (default account)
-  if (dingtalk.clientId || dingtalk.clientSecret) {
-    accountIds.push(DEFAULT_ACCOUNT_ID);
-  }
-
-  // Check accounts object
-  if (dingtalk.accounts) {
-    accountIds.push(...Object.keys(dingtalk.accounts));
-  }
-
-  return accountIds;
-}
-
-/**
- * Resolved DingTalk account with configuration status
- */
-export interface ResolvedDingTalkAccount extends DingTalkConfig {
-  accountId: string;
-  configured: boolean;
-}
-
-/**
- * Resolve a specific DingTalk account configuration
- */
-export function resolveDingTalkAccount(
-  cfg: OpenClawConfig,
-  accountId?: string | null,
-): ResolvedDingTalkAccount {
-  const id = accountId || DEFAULT_ACCOUNT_ID;
-  const dingtalk = cfg.channels?.dingtalk as DingTalkChannelConfig | undefined;
-
-  // If default account, return top-level config
-  if (id === DEFAULT_ACCOUNT_ID) {
-    const rawConfig: DingTalkConfig = {
-      clientId: dingtalk?.clientId ?? "",
-      clientSecret: dingtalk?.clientSecret ?? "",
-      name: dingtalk?.name,
-      enabled: dingtalk?.enabled,
-      dmPolicy: dingtalk?.dmPolicy,
-      groupPolicy: dingtalk?.groupPolicy,
-      allowFrom: dingtalk?.allowFrom,
-      groupAllowFrom: dingtalk?.groupAllowFrom,
-      displayNameResolution: dingtalk?.displayNameResolution,
-      contextVisibility: dingtalk?.contextVisibility,
-      journalTTLDays: dingtalk?.journalTTLDays,
-      ackReaction: dingtalk?.ackReaction,
-      debug: dingtalk?.debug,
-      messageType: dingtalk?.messageType,
-      cardTemplateId: dingtalk?.cardTemplateId,
-      cardTemplateKey: dingtalk?.cardTemplateKey,
-      groups: dingtalk?.groups,
-      accounts: dingtalk?.accounts,
-      maxConnectionAttempts: dingtalk?.maxConnectionAttempts,
-      initialReconnectDelay: dingtalk?.initialReconnectDelay,
-      maxReconnectDelay: dingtalk?.maxReconnectDelay,
-      reconnectJitter: dingtalk?.reconnectJitter,
-      maxReconnectCycles: dingtalk?.maxReconnectCycles,
-      reconnectDeadlineMs: dingtalk?.reconnectDeadlineMs,
-      useConnectionManager: dingtalk?.useConnectionManager,
-      mediaMaxMb: dingtalk?.mediaMaxMb,
-      keepAlive: dingtalk?.keepAlive,
-      bypassProxyForSend: dingtalk?.bypassProxyForSend,
-      proactivePermissionHint: dingtalk?.proactivePermissionHint,
-      cardStreamingMode: dingtalk?.cardStreamingMode,
-      cardRealTimeStream: dingtalk?.cardRealTimeStream,
-      cardStreamInterval: dingtalk?.cardStreamInterval,
-      aicardDegradeMs: dingtalk?.aicardDegradeMs,
-      learningEnabled: dingtalk?.learningEnabled,
-      learningAutoApply: dingtalk?.learningAutoApply,
-      learningNoteTtlMs: dingtalk?.learningNoteTtlMs,
-      convertMarkdownTables: dingtalk?.convertMarkdownTables,
-      cardAtSender: dingtalk?.cardAtSender,
-    };
-    const config = stripRemovedLegacyFieldsFromPublicAccount(rawConfig);
-    return {
-      ...config,
-      accountId: id,
-      configured: Boolean(config.clientId && config.clientSecret),
-    };
-  }
-
-  // If named account, merge channel-level defaults with account-level overrides
-  const accountConfig = dingtalk?.accounts?.[id];
-  if (accountConfig) {
-    const merged = mergeAccountWithDefaults(
-      dingtalk as DingTalkConfig,
-      accountConfig,
-    );
-    const publicMerged = stripRemovedLegacyFieldsFromPublicAccount(merged);
-    return {
-      ...publicMerged,
-      accountId: id,
-      configured: Boolean(merged.clientId && merged.clientSecret),
-    };
-  }
-
-  // Account doesn't exist, return empty config
-  return {
-    clientId: "",
-    clientSecret: "",
-    accountId: id,
-    configured: false,
-  };
-}
